% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_nested_concord.R
\name{CEL_lm_nested}
\alias{CEL_lm_nested}
\title{Run the nested algorithm for linear regression composite tests.}
\usage{
CEL_lm_nested(beta_test, X, y, F_reparam, s_hat_reparam, gamma_init = NULL,
  outer_eps = 1e-08, outer_maxit = 1000, inner_opt_type = c("owen",
  "LBFGS"), inner_owen_arg = optim_owen_inner_control(),
  inner_lbfgs_arg = list(invisible = 1),
  beta_newton_arg = optim_newton_control(), outer_tol_type = c("fval",
  "gval"), verbose = F)
}
\arguments{
\item{beta_test}{a vector of candidate test values.}

\item{X}{a design matrix with observations in rows and variables in columns.}

\item{y}{a vector of the responses.}

\item{F_reparam}{the \code{F} matrix for the affine reparameterization.}

\item{s_hat_reparam}{the \code{s} vector for the affine reparameterization.}

\item{gamma_init}{a vector of initial values for the dual, or a \code{"random"} to specify sampling from \code{\link[stats]{rnorm}}. Default is zero vector.}

\item{outer_eps}{absolute tolerance required for outer loop convergence.}

\item{outer_maxit}{maximum number of outer loop iterations.}

\item{inner_opt_type}{optimization type for the inner loop.}

\item{inner_owen_arg}{control arguments passed to the dual formulation (\code{owen}) optimization for the inner loop; see \code{\link{optim_owen_inner_control}}.}

\item{inner_lbfgs_arg}{a list of arguments passed to \code{\link[lbfgs]{lbfgs}}.}

\item{beta_newton_arg}{control arguments passed to the damped Newton optimization for the inner loop; see \code{\link{optim_newton_control}}.}

\item{outer_tol_type}{the type of tolerance checking for the outer loop.}

\item{verbose}{a boolean to allow console output.}
}
\description{
Run the nested algorithm for linear regression composite tests.
}
