% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owenEL.R
\name{emplik_concord}
\alias{emplik_concord}
\title{Use the dual formulation to solve the inner loop.}
\usage{
emplik_concord(z, lam, eps = 1/nrow(z), M = Inf, thresh = 1e-16,
  itermax = 150, verbose = FALSE)
}
\arguments{
\item{z}{a matrix of the score equations with observations in rows and variables in columns.}

\item{lam}{a vector of initial values for the dual parameter.}

\item{eps}{lower cutoff for logarithm approximation.}

\item{M}{upper cutoff for logarithm approximation.}

\item{thresh}{convergence threshold.}

\item{itermax}{maximum number of iterations.}

\item{verbose}{print console output.}
}
\description{
Use the dual formulation to solve the inner loop.
}
