% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_arg.R
\name{optim_mirror_control}
\alias{optim_mirror_control}
\title{Control parameters for mirror descent}
\usage{
optim_mirror_control(maxit = 1000, eps = 1e-08, tol_type = c("fval",
  "wts", "logelr"), vary_step = F, line_search = F, ls_beta = 0.75,
  ls_eps = 1e-10, ls_maxit = 100)
}
\arguments{
\item{maxit}{the maximum number of iterations.}

\item{eps}{the maximum value of the tolerance before declaring convergence.}

\item{tol_type}{the type of tolerance checking.}

\item{vary_step}{for a given step size, a boolean indicating whether it should be decreased at every iteration.}

\item{line_search}{a boolean to turn on line search - it is NOT recommended to turn this off.}

\item{ls_beta}{by how much should the step-size increase or decrease in each line search iteration?}

\item{ls_eps}{you can miss the line search check by this much.}

\item{ls_maxit}{the maximum number of line search iterations.}
}
\description{
Control parameters for mirror descent
}
