% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_arg.R
\name{optim_newton_control}
\alias{optim_newton_control}
\title{Control parameters for damped Newton}
\usage{
optim_newton_control(ls_beta = 0.7, ls_eps = 1e-12, ls_c1 = 0.25,
  ls_maxit = 100, mod_hess = F)
}
\arguments{
\item{ls_beta}{by how much should the step-size increase or decrease in each line search iteration?}

\item{ls_eps}{you can miss the line search check by this much.}

\item{ls_c1}{the fraction of the decrease in f predicted by linear extrapolation that we will accept.}

\item{ls_maxit}{the maximum number of line search iterations.}

\item{mod_hess}{the Hessian to be positive definite - recommended to leave this \code{FALSE}.}
}
\description{
Control parameters for damped Newton
}
