% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_arg.R
\name{optim_outlier_control}
\alias{optim_outlier_control}
\title{Control parameters for the proximal gradient descent update for outlier detection}
\usage{
optim_outlier_control(maxit = 2000, eps = 1e-05, accelerate = T,
  line_search = T, ls_beta = 0.75, ls_eps = 1e-12, ls_maxit = 100)
}
\arguments{
\item{maxit}{the maximum number of iterations.}

\item{eps}{the maximum value of the tolerance before declaring convergence.}

\item{accelerate}{a boolean to turn on acceleration.}

\item{line_search}{a boolean to turn on line search - it is NOT recommended to turn this off.}

\item{ls_beta}{by how much should the step-size increase or decrease in each line search iteration?}

\item{ls_eps}{you can miss the line search check by this much.}

\item{ls_maxit}{the maximum number of line search iterations.}
}
\description{
Control parameters for the proximal gradient descent update for outlier detection
}
