% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_arg.R
\name{optim_owen_inner_control}
\alias{optim_owen_inner_control}
\title{Control parameters for the inner loop dual formulation}
\usage{
optim_owen_inner_control(thresh = 1e-16, itermax = 150, verbose = F)
}
\arguments{
\item{thresh}{convergence threshold.}

\item{itermax}{maximum number of iterations.}

\item{verbose}{print console output.}
}
\description{
Control parameters for the inner loop dual formulation
}
