% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_arg.R
\name{optim_prox_control}
\alias{optim_prox_control}
\title{Control parameters for proximal gradient descent}
\usage{
optim_prox_control(tol_type = c("par", "fval"), step_size = 1,
  accelerate = F, line_search = T, ls_beta = 0.75, ls_eps = 1e-10,
  ls_maxit = 50, maxit = 1000, eps = 1e-06)
}
\arguments{
\item{tol_type}{the type of tolerance checking.}

\item{step_size}{initial step size.}

\item{accelerate}{a boolean to turn on acceleration.}

\item{line_search}{a boolean to turn on line search - it is NOT recommended to turn this off.}

\item{ls_beta}{by how much should the step-size increase or decrease in each line search iteration?}

\item{ls_eps}{you can miss the line search check by this much.}

\item{ls_maxit}{the maximum number of line search iterations.}

\item{maxit}{the maximum number of iterations.}

\item{eps}{the maximum value of the tolerance before declaring convergence.}
}
\description{
Control parameters for proximal gradient descent
}
