% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_operators.R
\name{prox_affine}
\alias{prox_affine}
\title{Proximal operator for affine reparameterization}
\usage{
prox_affine(inp, T_reparam, alpha_reparam)
}
\arguments{
\item{inp}{a vector.}

\item{T_reparam}{a matrix.}

\item{alpha_reparam}{a vector or double.}
}
\description{
Proximal operator for affine reparameterization
}
