% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_operators.R
\name{prox_inequality}
\alias{prox_inequality}
\title{Proximal operator for inequality testing}
\usage{
prox_inequality(inp, test_idx, test_val, dir = c("gt", "lt"))
}
\arguments{
\item{inp}{a vector.}

\item{test_idx}{the index for the values to be tested.}

\item{test_val}{the values for the test.}

\item{dir}{the direction of the inequality; greater than (\code{gt}) or less than (\code{lt}).}
}
\description{
Proximal operator for inequality testing
}
