% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_operators.R
\name{prox_interval}
\alias{prox_interval}
\title{Proximal operator for interval testing}
\usage{
prox_interval(inp, test_idx, left_val, right_val)
}
\arguments{
\item{inp}{a vector.}

\item{test_idx}{index for the values to be tested.}

\item{left_val}{a vector of values for the left interval. If the dimension does not agree with \code{test_idx}, the test is repeated for all indices in \code{test_idx}.}

\item{right_val}{the values for the right interval. If the dimension does not agree with \code{test_idx}, the test is repeated for all indices in \code{test_idx}.}
}
\description{
Proximal operator for interval testing
}
