% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_REL.R
\name{REL_fixed}
\alias{REL_fixed}
\title{Compute REL for a fixed estimate.}
\usage{
REL_fixed(Z_mat, q = NULL, outlier_loop_arg = list(), wts_init = NULL,
  vary_penalty = c("RB", "RB2", "none"), reuse_delta = F, RB_mu = 10,
  RB_tau = 2, RB2_ksi = 2, outer_eps = 1e-08, outer_rel_eps = 1e-04,
  dual_step = 2, outer_maxit = 1000, wts_beta_rep = 1,
  outer_tol_type = c("primal_dual", "fval", "gamma", "primal"),
  mirror_arg = optim_mirror_control(line_search = T), verbose = F)
}
\arguments{
\item{Z_mat}{a matrix of the score equations with observations in rows and variables in columns.}

\item{q}{an integer specifying the number of detected outliers. Default is \code{floor(.1 * n)}.}

\item{outlier_loop_arg}{control arguments passed to the \code{delta} optimization; see \code{\link{optim_outlier_control}}.}

\item{wts_init}{an initial guess for the weights vector. Default is \code{1/n}.}

\item{vary_penalty}{a string specifying the type of penalty variation.}

\item{reuse_delta}{use the last estimate of \code{delta} in the current outer loop iteration.}

\item{RB_mu}{the \code{mu} parameter for the residual balancing scheme.}

\item{RB_tau}{the \code{tau} parameter for the residual balancing scheme.}

\item{RB2_ksi}{the \code{ksi} parameter for the \code{RB2} scheme.}

\item{outer_eps}{absolute tolerance required for outer loop convergence.}

\item{outer_rel_eps}{relative tolerance required for outer loop convergence.}

\item{dual_step}{initial penalty parameter.}

\item{outer_maxit}{maximum number of outer loop iterations.}

\item{wts_beta_rep}{the number of repetitions of the block coordinate descent update within each outer loop iteration.}

\item{outer_tol_type}{the type of tolerance check for the outer loop.}

\item{mirror_arg}{control arguments passed to the mirror descent optimization; see \code{\link{optim_mirror_control}}.}

\item{verbose}{a boolean to allow console output.}
}
\value{
A list with estimates and convergence information.
}
\description{
Compute REL for a fixed estimate.
}
