% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_REL.R
\name{REL_mean_CI}
\alias{REL_mean_CI}
\title{Profile a confidence interval for a one-dimensional mean using either REL or EL}
\usage{
REL_mean_CI(data_mat, q = NULL, conf_level = 0.05, test_thresh = "chisq",
  upper_break = NULL, lower_break = 1e-06, upper_divide = 5,
  upper_increase = 0.1, left_algorithm = c("REL", "EL"),
  right_algorithm = c("REL", "EL"), ..., verbose = F)
}
\arguments{
\item{data_mat}{a matrix of the data with observations in rows and variables in columns.}

\item{q}{an integer specifying the number of detected outliers. Default is \code{floor(.1 * n)}.}

\item{conf_level}{the confidence level.}

\item{test_thresh}{a string specifying the calibration type. Only the chi-squared approximation is currently implemented.}

\item{upper_break}{the initial guess for the upper limit for root finding. Default uses a value based on the magnitude of the MLE.}

\item{lower_break}{the initial guess for the lower limit for root finding.}

\item{upper_divide}{divide the MLE by this much for \code{upper_break}.}

\item{upper_increase}{increase the initial guess for the upper limit by this amount.}

\item{left_algorithm}{the algorithm for the left confidence limit.}

\item{right_algorithm}{the algorithm for the right confidence limit.}

\item{...}{additional arguments passed to \code{REL_fixed}.}

\item{verbose}{a boolean to allow console output.}
}
\value{
A list with confidence interval attributes.
}
\description{
Profile a confidence interval for a one-dimensional mean using either REL or EL
}
