% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_reparam.R
\name{reparam_helper}
\alias{reparam_helper}
\title{Construct the variables for the affine reparameterization}
\usage{
reparam_helper(mean_init, fix_index, alpha_add)
}
\arguments{
\item{mean_init}{a vector of the MLE.}

\item{fix_index}{a vector containing the indices to be profiled.}

\item{alpha_add}{a vector of the same length as \code{fix_index} containing the adjustments away from the MLE.}
}
\value{
A list with the new candidate value for \code{beta} and the matrices/vectors required for the reparameterization.
}
\description{
Construct the variables for the affine reparameterization
}
