% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRICEP_profiler.R
\name{TRICEP_glm_beta_add}
\alias{TRICEP_glm_beta_add}
\title{Compute CEL by perturbing the MLE at a coefficient.}
\usage{
TRICEP_glm_beta_add(alpha_add, fix_index, X, y = NULL,
  family = c("gaussian", "binomial", "poisson", "mean"), mean_init = NULL,
  algorithm = c("TRICEP", "nested"), ...)
}
\arguments{
\item{alpha_add}{a vector with the perturbations.}

\item{fix_index}{a vector indicating the indices to be tested.}

\item{X}{a design matrix with observations in rows and variables in columns.}

\item{y}{a vector of the responses. Default is \code{NULL} in case \code{family = "mean"} (test a vector mean).}

\item{family}{either the GLM family or \code{"mean"} for a vector mean.}

\item{mean_init}{the MLE for the coefficients.}

\item{algorithm}{a string indicating the algorithm to be used. \code{"nested"} is only compatible when \code{family = "gaussian"}.}

\item{...}{additional arguments passed to \code{\link{TRICEP_glm}}.}
}
\value{
A list with the output from \code{\link{TRICEP_glm}}.
}
\description{
Compute CEL by perturbing the MLE at a coefficient.
}
