% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRICEP_profiler.R
\name{TRICEP_glm_beta_boot}
\alias{TRICEP_glm_beta_boot}
\title{Bootstrap TRICEP/BICEP for a coefficient, starting at the MLE.}
\usage{
TRICEP_glm_beta_boot(fix_index, X, y = NULL, family = c("gaussian",
  "binomial", "poisson"), B = 200, prob = 0.95, mean_init = NULL,
  algorithm = c("TRICEP", "nested"), parallel = T, seed = 123, ...)
}
\arguments{
\item{fix_index}{a vector indicating the indices to be boostrapped.}

\item{X}{a design matrix with observations in rows and variables in columns.}

\item{y}{a vector of the responses. Default is \code{NULL} in case \code{family = "mean"} (test a vector mean).}

\item{family}{either the GLM family or \code{"mean"} for a vector mean.}

\item{B}{the number of boostrap replications.}

\item{prob}{the coverage probability.}

\item{mean_init}{the MLE for the coefficients.}

\item{algorithm}{a string indicating the algorithm to be used. \code{"nested"} is only compatible when \code{family = "gaussian"}.}

\item{parallel}{a boolean indicating whether each replicate should be parallelized. A backend for \code{\link[foreach]{foreach}} must be registered if this is \code{T}.}

\item{seed}{an integer for setting the seed.}

\item{...}{additional arguments passed to \code{\link{TRICEP_glm}}.}
}
\value{
A list with the bootstrapped test statistic and the replicates.
}
\description{
Bootstrap TRICEP/BICEP for a coefficient, starting at the MLE.
}
