% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRICEP_profiler.R
\name{TRICEP_glm_beta_profiler}
\alias{TRICEP_glm_beta_profiler}
\title{Profile a CEL confidence interval for one component of the coefficient vector.}
\usage{
TRICEP_glm_beta_profiler(fix_index, X, y = NULL, family = c("gaussian",
  "binomial", "poisson", "mean"), conf_level = 0.05, test_thresh = "chisq",
  upper_break = NULL, lower_break = 1e-06, upper_divide = 2,
  upper_increase = 0.1, algorithm = c("TRICEP", "nested"), ...,
  verbose = F)
}
\arguments{
\item{fix_index}{a vector indicating the indices to be tested.}

\item{X}{a design matrix with observations in rows and variables in columns.}

\item{y}{a vector of the responses. Default is \code{NULL} in case \code{family = "mean"} (test a vector mean).}

\item{family}{either the GLM family or \code{"mean"} for a vector mean.}

\item{conf_level}{the confidence level.}

\item{test_thresh}{either a string specifying the chi-square approximation or a double (such as a statistic calculated from the bootstrap).}

\item{upper_break}{the initial guess for the upper limit for root finding. Default uses a value based on the magnitude of the MLE.}

\item{lower_break}{the initial guess for the lower limit for root finding.}

\item{upper_divide}{divide the MLE by this much for \code{upper_break}.}

\item{upper_increase}{increase the initial guess for the upper limit by this amount.}

\item{algorithm}{a string indicating the algorithm to be used. \code{"nested"} is only compatible when \code{family = "gaussian"}.}

\item{...}{additional arguments passed to \code{\link{TRICEP_glm}}.}

\item{verbose}{a boolean to allow console output.}
}
\value{
A list with confidence interval attributes.
}
\description{
Profile a CEL confidence interval for one component of the coefficient vector.
}
