% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRICEP_profiler.R
\name{TRICEP_glm_boot}
\alias{TRICEP_glm_boot}
\title{Bootstrap TRICEP/BICEP for general composite tests using a proximal operator}
\usage{
TRICEP_glm_boot(X, y = NULL, prox_fun = prox_inequality,
  prox_fun_arg = list(test_idx = 1, test_val = 0), family = c("gaussian",
  "binomial", "poisson", "mean"), ..., beta_init = NULL, B = 200,
  prob = 0.95, parallel = T, seed = 123)
}
\arguments{
\item{X}{a design matrix with observations in rows and variables in columns.}

\item{y}{a vector of the responses. Default is \code{NULL} in case \code{family = "mean"} (test a vector mean).}

\item{prox_fun}{a proximal operator. This should be an \code{R} object.}

\item{prox_fun_arg}{a list of arguments passed to \code{prox_fun}.}

\item{family}{either the GLM family or \code{"mean"} for a vector mean.}

\item{...}{additional arguments passed to \code{\link{TRICEP_glm}}.}

\item{beta_init}{an initial guess for the \code{beta} vector.}

\item{B}{the number of boostrap replications.}

\item{prob}{the coverage probability.}

\item{parallel}{a boolean indicating whether each replicate should be parallelized. A backend for \code{\link[foreach]{foreach}} must be registered if this is \code{T}.}

\item{seed}{an integer for setting the seed.}
}
\value{
A list with the bootstrapped test statistic, the p value, and the replicates.
}
\description{
Bootstrap TRICEP/BICEP for general composite tests using a proximal operator
}
