% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRICEP.R
\name{TRICEP_mean}
\alias{TRICEP_mean}
\title{Run TRICEP/BICEP for a vector mean.}
\usage{
TRICEP_mean(beta_test, X, F_reparam = NULL, s_hat_reparam = NULL,
  prox_fun = NULL, detect_outlier = F, q = NULL, reuse_delta = F,
  prox_fun_arg = list(), wts_init = NULL, vary_penalty = c("RB", "RB2",
  "none"), RB_mu = 10, RB_tau = 2, RB2_ksi = 2, outer_eps = 1e-08,
  outer_rel_eps = 1e-04, dual_step = 2, outer_maxit = 1000,
  wts_beta_rep = 1, outer_tol_type = c("primal_dual", "fval", "primal"),
  outlier_loop_arg = optim_outlier_control(),
  mirror_arg = optim_mirror_control(line_search = T),
  prox_optim_arg = optim_prox_control(), verbose = F)
}
\arguments{
\item{beta_test}{a vector of candidate test values.}

\item{X}{a design matrix with observations in rows and variables in columns.}

\item{F_reparam}{the \code{F} matrix for the affine reparameterization.}

\item{s_hat_reparam}{the \code{s} vector for the affine reparameterization.}

\item{prox_fun}{a proximal operator. This should be an \code{R} object.}

\item{detect_outlier}{a boolean indicating whether an REL update should be run to detect outliers.}

\item{q}{the number of detected outliers (only valid if \code{detect_outlier = T}). The default is \code{floor(.05 * n)}.}

\item{reuse_delta}{use the last estimate of \code{delta} in the current outer loop iteration.}

\item{prox_fun_arg}{a list of arguments passed to \code{prox_fun}.}

\item{wts_init}{an initial guess for the weights vector. Default is \code{1/n}.}

\item{vary_penalty}{a string specifying the type of penalty variation.}

\item{RB_mu}{the \code{mu} parameter for the residual balancing scheme.}

\item{RB_tau}{the \code{tau} parameter for the residual balancing scheme.}

\item{RB2_ksi}{the \code{ksi} parameter for the \code{RB2} scheme.}

\item{outer_eps}{absolute tolerance required for outer loop convergence.}

\item{outer_rel_eps}{relative tolerance required for outer loop convergence.}

\item{dual_step}{initial penalty parameter.}

\item{outer_maxit}{maximum number of outer loop iterations.}

\item{wts_beta_rep}{the number of repetitions of the block coordinate descent update within each outer loop iteration.}

\item{outer_tol_type}{the type of tolerance check for the outer loop.}

\item{outlier_loop_arg}{control arguments passed to the \code{delta} optimization; see \code{\link{optim_outlier_control}}.}

\item{mirror_arg}{control arguments passed to the mirror descent optimization; see \code{\link{optim_mirror_control}}.}

\item{prox_optim_arg}{a list of control arguments passed to the proximal gradient descent update for \code{prox_fun}; see \code{\link{optim_prox_control}}.}

\item{verbose}{a boolean to allow console output.}
}
\description{
Run TRICEP/BICEP for a vector mean.
}
